Change rate limit datatype to uint32 to not lose precision when casting
We don't want to do weird things for images bigger 2.147GB and using
uint64 is not that trivial in lua which the rate limiting module uses

+++ internal/api/utils.go
@@ -29,7 +29,7 @@ import (
 	"github.com/sapcc/keppel/internal/keppel"
 )
 
-func CheckRateLimit(r *http.Request, rle *keppel.RateLimitEngine, account keppel.Account, authz *auth.Authorization, action keppel.RateLimitedAction, amount uint64) error {
+func CheckRateLimit(r *http.Request, rle *keppel.RateLimitEngine, account keppel.Account, authz *auth.Authorization, action keppel.RateLimitedAction, amount uint32) error {
 	//rate-limiting is optional
 	if rle == nil {
 		return nil
